# src/backend/rest/api/swagger/swagger.yaml
swagger: '2.0'
info:
  version: '1.0'
  title: FridoLean.io
host: localhost:3001
basePath: /
schemes:
  - http
  - https
consumes:
  - application/json
produces:
  - application/json
...
paths:
...
  /api/v1/projects:
    x-swagger-router-controller: project
    get:
      tags:
        - project
      description: Returns list of publicly visible projects
      operationId: getPublicProjects
      parameters:
        - name: Authorization
          in: header
          type: string
          required: false
          description: Bearer token
        - name: title
          in: query
          description: Title of an public project
          type: string
        - name: start
          in: query
          description: The number of projects to skip. Default 0.
          required: false
          type: integer
        - name: count
          in: query
          description: The number of projects returned (at the maximum) Default 20.
          type: integer
          required: false
      responses:
        '200':
          description: Success
          schema:
            type: array
            items:
              $ref: '#/definitions/Project'
        default:
          description: Error
          schema:
            $ref: '#/definitions/ErrorResponse'
    post:
      tags:
        - project
      description: Creates new project for a user
      operationId: createProject
      parameters:
        - name: Authorization
          in: header
          type: string
          required: true
          description: Bearer token
        - name: data
          required: true
          description: Metadata of the new project
          in: body
          schema:
            $ref: '#/definitions/CreateProject'
      responses:
        '200':
          description: Success
          schema:
            $ref: '#/definitions/ProjectMeta'
        '400':
          description: Bad Request
          schema:
            $ref: '#/definitions/ErrorResponse'
        '401':
          description: Unauthorized
          schema:
            $ref: '#/definitions/ErrorResponse'
        '500':
          description: Internal Server Error
          schema:
            $ref: '#/definitions/ErrorResponse'
        default:
          description: Error
          schema:
            $ref: '#/definitions/ErrorResponse'
  '/api/v1/projects/{projectID}':
    x-swagger-router-controller: project
    get:
      tags:
        - project
      description: Returns information about specific project
      operationId: getProjectInfo
      parameters:
        - name: Authorization
          in: header
          type: string
          required: true
          description: Bearer token
        - name: projectID
          in: path
          required: true
          type: string
          format: uuid
      responses:
        '200':
          description: Success
          schema:
            $ref: '#/definitions/Project'
        '400':
          description: Bad Request
          schema:
            $ref: '#/definitions/ErrorResponse'
        '401':
          description: Unauthorized
          schema:
            $ref: '#/definitions/ErrorResponse'
        '500':
          description: Internal Server Error
          schema:
            $ref: '#/definitions/ErrorResponse'
        default:
          description: Error
          schema:
            $ref: '#/definitions/ErrorResponse'
    put:
      tags:
        - project
      description: Updates data of a project
      operationId: updateProject
      parameters:
        - name: Authorization
          in: header
          type: string
          required: true
          description: Bearer token
        - name: projectID
          in: path
          required: true
          type: string
          format: uuid
        - name: data
          required: true
          description: Updated Metadata of the project
          in: body
          schema:
            $ref: '#/definitions/CreateProject'
      responses:
        '200':
          description: Success
        '400':
          description: Bad Request
          schema:
            $ref: '#/definitions/ErrorResponse'
        '401':
          description: Unauthorized
          schema:
            $ref: '#/definitions/ErrorResponse'
        '500':
          description: Internal Server Error
          schema:
            $ref: '#/definitions/ErrorResponse'
        default:
          description: Error
          schema:
            $ref: '#/definitions/ErrorResponse'
    delete:
      tags:
        - project
      description: Delete specific project
      operationId: deleteProject
      parameters:
        - name: Authorization
          in: header
          type: string
          required: true
          description: Bearer token
        - name: projectID
          in: path
          required: true
          type: string
          format: uuid
      responses:
        '200':
          description: Success
        '400':
          description: Bad Request
          schema:
            $ref: '#/definitions/ErrorResponse'
        '401':
          description: Unauthorized
          schema:
            $ref: '#/definitions/ErrorResponse'
        '403':
          description: Forbidden
          schema:
            $ref: '#/definitions/ErrorResponse'
        '500':
          description: Internal Server Error
          schema:
            $ref: '#/definitions/ErrorResponse'
        default:
          description: Error
          schema:
            $ref: '#/definitions/ErrorResponse'
  '/api/v1/projects/{projectID}/collaborators':
    x-swagger-router-controller: collaborator
    post:
      tags:
        - project
      description: Add collaborator to project
      operationId: addCollaborator
      parameters:
        - name: Authorization
          in: header
          type: string
          required: true
          description: Bearer token
        - name: projectID
          in: path
          required: true
          type: string
          format: uuid
        - name: data
          required: true
          description: Information about new collaborator
          in: body
          schema:
            $ref: '#/definitions/CollaboratorData'
      responses:
        '200':
          description: Success
        '400':
          description: Bad Request
          schema:
            $ref: '#/definitions/ErrorResponse'
        '401':
          description: Unauthorized
          schema:
            $ref: '#/definitions/ErrorResponse'
        '403':
          description: Forbidden
          schema:
            $ref: '#/definitions/ErrorResponse'
        '500':
          description: Internal Server Error
          schema:
            $ref: '#/definitions/ErrorResponse'
        default:
          description: Error
          schema:
            $ref: '#/definitions/ErrorResponse'
    put:
      tags:
        - project
      description: Update collaborator in project
      operationId: updateCollaborator
      parameters:
        - name: Authorization
          in: header
          type: string
          required: true
          description: Bearer token
        - name: projectID
          in: path
          required: true
          type: string
          format: uuid
        - name: data
          required: true
          description: Updated information about collaborator
          in: body
          schema:
            $ref: '#/definitions/CollaboratorData'
      responses:
        '200':
          description: Success
        '400':
          description: Bad Request
          schema:
            $ref: '#/definitions/ErrorResponse'
        '401':
          description: Unauthorized
          schema:
            $ref: '#/definitions/ErrorResponse'
        '403':
          description: Forbidden
          schema:
            $ref: '#/definitions/ErrorResponse'
        '500':
          description: Internal Server Error
          schema:
            $ref: '#/definitions/ErrorResponse'
        default:
          description: Error
          schema:
            $ref: '#/definitions/ErrorResponse'
    delete:
      tags:
        - project
      description: Delete collaborator from project
      operationId: deleteCollaborator
      parameters:
        - name: Authorization
          in: header
          type: string
          required: true
          description: Bearer token
        - name: projectID
          in: path
          required: true
          type: string
          format: uuid
        - name: data
          required: true
          description: Identification of collaborator to remove
          in: body
          schema:
            properties:
              email:
                type: string
      responses:
        '200':
          description: Success
        '400':
          description: Bad Request
          schema:
            $ref: '#/definitions/ErrorResponse'
        '401':
          description: Unauthorized
          schema:
            $ref: '#/definitions/ErrorResponse'
        '403':
          description: Forbidden
          schema:
            $ref: '#/definitions/ErrorResponse'
        '500':
          description: Internal Server Error
          schema:
            $ref: '#/definitions/ErrorResponse'
        default:
          description: Error
          schema:
            $ref: '#/definitions/ErrorResponse'
...
  '/api/v1/projects/{projectID}/canvas/{canvasID}/pdf':
    x-swagger-router-controller: pdf
    get:
      tags:
        - project
        - entry
        - canvas
        - pdf
      description: Renders a pdf file and sends it in base64 format
      operationId: sendPDF
      parameters:
        - name: Authorization
          in: header
          type: string
          required: false
          description: Bearer token
        - name: projectID
          in: path
          required: true
          type: string
          format: uuid
        - name: canvasID
          in: path
          required: true
          type: string
          format: uuid
        - name: tag
          in: query
          required: false
          type: string
          format: uuid
        - name: size
          in: query
          default: default
          type: string
          enum:
            - default
            - A4
            - A3
            - A2
            - A1
      responses:
        '200':
          description: Success
          schema:
            $ref: '#/definitions/PdfData'
        '400':
          description: Bad Request
          schema:
            $ref: '#/definitions/ErrorResponse'
        '401':
          description: Unauthorized
          schema:
            $ref: '#/definitions/ErrorResponse'
        '403':
          description: Forbidden
          schema:
            $ref: '#/definitions/ErrorResponse'
        '500':
          description: Internal Server Error
          schema:
            $ref: '#/definitions/ErrorResponse'
        default:
          description: Error
          schema:
            $ref: '#/definitions/ErrorResponse'
...
definitions:
...
  persistent:
    properties:
      canvases:
        type: array
        items:
          $ref: '#/definitions/Canvas'
  CreateProject:
    properties:
      title:
        type: string
        default: Untitled
        description: Title of the project
      description:
        description: Description of the project
        type: string
      visibility:
        type: string
        description: PUBLIC || PRIVATE
  Project:
    properties:
      _id:
        type: string
        format: uuid
      title:
        type: string
      visibility:
        type: string
        description: visibility tag like PUBLIC or PRIVATE
      lastEdited:
        type: integer
        format: int64
        description: 'Milliseconds since Unix epoch (Jan 1, 1970)'
      description:
        type: string
      members:
        type: array
        items:
          properties:
            userID:
              type: string
            email:
              type: string
            rights:
              type: string
              description: READ || EDIT
      timeline:
        type: array
        items:
          properties:
            timestamp:
              type: number
            tag:
              type: string
            value:
              $ref: '#/definitions/persistent'
      persistent:
        $ref: '#/definitions/persistent'
  ProjectMeta:
    properties:
      _id:
        type: string
        format: uuid
      title:
        type: string
      visibility:
        type: string
        description: PUBLIC || PRIVATE
      lastEdited:
        type: integer
        format: int64
        description: 'Milliseconds since Unix epoch (Jan 1, 1970)'
      description:
        type: string
      members:
        type: array
        items:
          properties:
            userID:
              type: string
            email:
              type: string
            rights:
              type: string
...
  CollaboratorData:
    properties:
      email:
        type: string
        format: email
      rights:
        type: string
        description: READ || EDIT
...
  Entry:
    properties:
      _id:
        type: string
      type:
        type: string
      content:
        $ref: '#/definitions/Content'
  Content:
    properties:
      title:
        type: string
      text:
        type: string
      reference:
        type: string
      target:
        type: string
  BuildingBlock:
    properties:
      _id:
        type: string
        format: uuid
      title:
        type: string
        description: |
          BuisinessModelCanvas: [
            KeyPartners,
            KeyActivities,
            KeyResources,
            CustomerRelation,
            Channels,
            CustomerSegments,
            CostStructure,
            RevenueStreams,
            ValueProposition
          ],                                    

          ValuePropositionCanvas: [      
            ProductsAndServices,
            GainCreators,
            PainRelievers,
            Gains,
            Pains ,                                
            CustomerJobs
          ],                            

          LeanCanvas: [
            Problem,                                  
            Solution,                                   
            KeyMetrics,                                      
            UniqueValueProposition,                                                   
            UnfairAdvantages,                                           
            Channels,                                   
            CustomerSegments,                                           
            CostStructure,                                         
            RevenueStreams
          ],                                 

          CustomerJourney: [
            Advertisement,                                        
            PreSocialMedia,                                              
            PreWordOfMouth,                                               
            Expectations,                                       
            PastExperiences,                                           
            ServiceJourney,                                          
            Experiences,                                      
            RelationshipMenage,                                                  
            PostSocialMedia,                                               
            PostWordOfMouth,                                                
            DisSatisfaction
          ]
      buildingBlockType:
        type: string
      layoutEntry:
        properties:
          _id:
            type: string
            format: uuid
          x:
            type: number
          'y':
            type: number
          width:
            type: number
          height:
            type: number
      entries:
        type: array
        items:
          $ref: '#/definitions/Entry'
...
  Canvas:
    properties:
      _id:
        type: string
        format: uuid
      title:
        type: string
      canvasType:
        type: string
        description: BUSINESS_MODEL || LEAN || VALUE_PROPOSITION || CUSTOMER_JOURNEY
      lastEdited:
        type: integer
        format: int64
        description: 'Milliseconds since Unix epoch (Jan 1, 1970)'
      buildingBlocks:
        type: array
        items:
          $ref: '#/definitions/BuildingBlock'
...
  PdfData:
    properties:
      pdf:
        type: string
        description: contains all data of the rendered pdf in base64 format
...
  ErrorResponse:
    required:
      - message
    properties:
      message:
        type: string
...